Added ability to specify which fields appear in JSON dumps.
Sweepable objects now automatically add user attributes to a bookkeeping
structure based on the type and name of the attribute, and only these
attributes are included in the JSON dump. This fixes the issue where
array sizes, word lengths, and other fields of the design sweep were not
in the dumps, since they are not sweepable parameters.