Add support for a source command.
The source command inserts commands from another file into the current
file, and the composite file is then executed as usual.  Unlike other
commands, the source command is "executed" at parse time and replaced by
the sourced commands, so it should never appear in a list of commands
after parsing is finished.

The filepath is specified as a string. If the filepath is relative, the
base directory is taken to be the directory containing the sourcing
file.

I also took the opportunity to restructure the command parsing modules.

Added test for this.